pyproject.toml
[build-system]
requires = ["setuptools>=62.0"]
build-backend = "setuptools.build_meta"

[project]
name = "restictool"
version = "0.7.0"
authors = [
  { name="Stanislav Meduna", email="stano@meduna.org" },
]
description = "A python wrapper for a dockerized restic tool"
readme = "README.rst"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "PyYAML",
    "docker",
    "schema",
    "setuptools",
    "python-dateutil",
    "prometheus-client"
]

[project.urls]
"Homepage" = "https://github.com/numo68/restictool"
"Bug Tracker" = "https://github.com/numo68/restictool/issues"

[project.scripts]
restictool-py = "restictool:main"

[tool.setuptools]
packages = ["restictool"]
package-dir = {"" = "src"}

[tool.pytest.ini_options]
pythonpath = [
  "src",
]
testpaths = [
    "tests",
]

[tool.bumpver]
current_version = "0.7.0"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "bump version {old_version} -> {new_version}"
tag_message = "{new_version}"
tag_scope = "default"
pre_commit_hook = ""
post_commit_hook = ""
commit = false
tag = false
push = false

[tool.bumpver.file_patterns]
"pyproject.toml" = [
    'current_version = "{version}"',
    'version = "{version}"',
]
"src/restictool/argument_parser.py" = [
    'self.version = "{version}"',
]
"docs/source/conf.py" = [
    'release = "{version}"',
]
